\name{simdata}
\alias{simdata}
\alias{simdata0}
\docType{data}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Simulated data}

\description{\code{simdata} is a simple test data set simulated from two clusters and consisting of 100 curves, with 50 curves belonging to one cluster and 50 to another. The test data set is a copy of the \code{simdata} of James and Sugar (2003).
}
%\usage{
%data(simdata)
%}
%- maybe also `usage' for other objects documented here.

\format{ \code{simdata} is a list of three vectors called as \code{x}, \code{curve} and \code{time}. 
\code{simdata0} is a reduced dataset with only six curves in each cluster.

} 


%\details{The data frame \code{simdata} relates to ......
%}

\source{
The data are from James and Sugar (2003).

}

%\examples{
%}
\references{
James, G.M., Sugar, C.A. (2003). Clustering for sparsely sampled functional data.  \emph{Journal of the American Statistical Association}, 98.462, 397--408.

}








